// interfaces/interfaces.ts
// Generated by https://quicktype.io

export interface PokemonApiResponse {
    count:    number;
    next?:     string;
    previous?: string;
    results:  Pokemons[];
}

export interface Pokemons {
    name : string;
    url  : string;
    id   : number;
    img  : string
};

type CardPokemons = Omit<Pokemons, 'url'>

export type PokeName = Pick< Pokemons, 'name' >

export interface CardPokemonsProps {
    pokemon:CardPokemons;
}